STY: Update test location.
