+++ artview/core/__init__.py
@@ -19,5 +19,5 @@ ARTview is based in the abstract Component class and shared Variables.
 """
 
 from . import common
-from .core import Variable, componentsList, Component, QtGui, QtCore, checkifmac
+from .core import Variable, componentsList, Component, QtGui, QtCore
 from .variable_choose import VariableChoose
